python-data-engineer/week12/exercises/specialized_storage.py
"""
Specialized storage exercises.
"""

from typing import Dict, Any, Optional, List, Union
import logging
from abc import ABC, abstractmethod
import numpy as np
import pandas as pd
from datetime import datetime
import deeplake
import networkx as nx
from pyspark.sql import SparkSession
from pyspark.sql.functions import *

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

class ImageStorage:
    """Image storage management using DeepLake."""
    
    def __init__(self, dataset_path: str):
        """Initialize storage."""
        self.dataset_path = dataset_path
        self.ds = None
        self._connect()
    
    def _connect(self) -> None:
        """Connect to DeepLake dataset."""
        try:
            self.ds = deeplake.load(self.dataset_path)
            
        except Exception:
            # Create new dataset
            self.ds = deeplake.empty(self.dataset_path)
            
            # Add tensors
            self.ds.create_tensor('images', htype='image')
            self.ds.create_tensor('metadata', htype='json')
            self.ds.create_tensor('embeddings', htype='embedding')
    
    def store_image(
        self,
        image: bytes,
        metadata: Dict[str, Any],
        embedding: Optional[np.ndarray] = None
    ) -> None:
        """
        Store image with metadata.
        
        Args:
            image: Image data
            metadata: Image metadata
            embedding: Image embedding
        """
        try:
            with self.ds:
                # Store image
                self.ds.images.append(image)
                
                # Store metadata
                self.ds.metadata.append(metadata)
                
                # Store embedding if provided
                if embedding is not None:
                    self.ds.embeddings.append(embedding)
            
            logger.info("Stored image successfully")
            
        except Exception as e:
            logger.error(f"Failed to store image: {e}")
            raise
    
    def search_similar(
        self,
        query_embedding: np.ndarray,
        k: int = 5
    ) -> List[Dict[str, Any]]:
        """
        Search similar images.
        
        Args:
            query_embedding: Query embedding
            k: Number of results
        
        Returns:
            List of similar images
        """
        try:
            # Search using embeddings
            results = self.ds.embeddings.search(
                query_embedding,
                k=k
            )
            
            # Get matching images and metadata
            similar = []
            for idx in results.ids:
                similar.append({
                    'image': self.ds.images[idx].numpy(),
                    'metadata': self.ds.metadata[idx].data()
                })
            
            return similar
            
        except Exception as e:
            logger.error(f"Failed to search images: {e}")
            raise

class TimeSeriesStorage:
    """Time series storage optimization."""
    
    def __init__(self, spark: SparkSession):
        """Initialize storage."""
        self.spark = spark
    
    def optimize_partitioning(
        self,
        df: Any,
        timestamp_col: str,
        partition_freq: str = 'month'
    ) -> Any:
        """
        Optimize time series partitioning.
        
        Args:
            df: Input DataFrame
            timestamp_col: Timestamp column
            partition_freq: Partition frequency
        
        Returns:
            Optimized DataFrame
        """
        try:
            # Add partition columns
            df = df.withColumn(
                'year',
                year(col(timestamp_col))
            )
            
            if partition_freq in ['month', 'day']:
                df = df.withColumn(
                    'month',
                    month(col(timestamp_col))
                )
                
            if partition_freq == 'day':
                df = df.withColumn(
                    'day',
                    dayofmonth(col(timestamp_col))
                )
            
            # Set partitioning
            partition_cols = ['year']
            if partition_freq in ['month', 'day']:
                partition_cols.append('month')
            if partition_freq == 'day':
                partition_cols.append('day')
            
            return df.repartition(*partition_cols)
            
        except Exception as e:
            logger.error(f"Failed to optimize partitioning: {e}")
            raise
    
    def downsample_series(
        self,
        df: Any,
        timestamp_col: str,
        value_col: str,
        window: str
    ) -> Any:
        """
        Downsample time series.
        
        Args:
            df: Input DataFrame
            timestamp_col: Timestamp column
            value_col: Value column
            window: Window size
        
        Returns:
            Downsampled DataFrame
        """
        try:
            return df.groupBy(
                window(col(timestamp_col), window)
            ).agg(
                avg(value_col).alias(f"{value_col}_avg"),
                min(value_col).alias(f"{value_col}_min"),
                max(value_col).alias(f"{value_col}_max")
            )
            
        except Exception as e:
            logger.error(f"Failed to downsample series: {e}")
            raise

class GraphStorage:
    """Graph data storage optimization."""
    
    def __init__(self, spark: SparkSession):
        """Initialize storage."""
        self.spark = spark
    
    def store_graph(
        self,
        nodes: Any,
        edges: Any,
        format: str = 'graphframes'
    ) -> Any:
        """
        Store graph data.
        
        Args:
            nodes: Node DataFrame
            edges: Edge DataFrame
            format: Storage format
        
        Returns:
            Graph object
        """
        try:
            if format == 'graphframes':
                from graphframes import GraphFrame
                return GraphFrame(nodes, edges)
            
            elif format == 'networkx':
                # Convert to NetworkX
                G = nx.Graph()
                
                # Add nodes
                for row in nodes.collect():
                    G.add_node(row['id'], **row.asDict())
                
                # Add edges
                for row in edges.collect():
                    G.add_edge(
                        row['src'],
                        row['dst'],
                        **row.asDict()
                    )
                
                return G
            
            else:
                raise ValueError(f"Invalid format: {format}")
            
        except Exception as e:
            logger.error(f"Failed to store graph: {e}")
            raise
    
    def optimize_traversal(
        self,
        graph: Any,
        traversal_type: str = 'bfs'
    ) -> Any:
        """
        Optimize graph traversal.
        
        Args:
            graph: Graph object
            traversal_type: Traversal type
        
        Returns:
            Optimized graph
        """
        try:
            if isinstance(graph, nx.Graph):
                if traversal_type == 'bfs':
                    # Optimize for BFS
                    return nx.optimize_graph_for_bfs(graph)
                
                elif traversal_type == 'dfs':
                    # Optimize for DFS
                    return nx.optimize_graph_for_dfs(graph)
                
            else:
                # GraphFrames optimization
                return graph.cache()
            
        except Exception as e:
            logger.error(f"Failed to optimize traversal: {e}")
            raise
    
    def analyze_structure(
        self,
        graph: Any
    ) -> Dict[str, Any]:
        """
        Analyze graph structure.
        
        Args:
            graph: Graph object
        
        Returns:
            Structure analysis
        """
        try:
            analysis = {}
            
            if isinstance(graph, nx.Graph):
                analysis.update({
                    'num_nodes': graph.number_of_nodes(),
                    'num_edges': graph.number_of_edges(),
                    'density': nx.density(graph),
                    'avg_degree': sum(dict(graph.degree()).values()) / graph.number_of_nodes(),
                    'connected_components': nx.number_connected_components(graph)
                })
            
            else:
                # GraphFrames analysis
                analysis.update({
                    'num_nodes': graph.vertices.count(),
                    'num_edges': graph.edges.count()
                })
            
            return analysis
            
        except Exception as e:
            logger.error(f"Failed to analyze structure: {e}")
            raise 